python/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
ftfy==6.0.3
idna==3.3
mailchecker==4.1.6
phonenumbers==8.12.39
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
PyYAML==6.0
requests==2.26.0
six==1.16.0
spotipy==2.19.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
wcwidth==0.2.5
xmltodict==0.12.0
